i_hate_save_games_with_all_my_heart_please_bring_me_back_fallout_3_save_games

+++ Assets/Scripts/Enemy/HurtBoxFatal.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [RequireComponent(typeof(BoxCollider2D))]
 public class HurtBoxFatal : MonoBehaviour
 {
-    float attackDamage = 100f;
+    int attackDamage = 100;
     float orientation = 1f;
     private GameObject player;
